# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "api",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/agents": {
      "get": {
        "tags": [
          "agents"
        ],
        "operationId": "list_agents",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Agent"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "agents"
        ],
        "operationId": "create_agent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent_id}": {
      "get": {
        "tags": [
          "agents"
        ],
        "operationId": "get_agent",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "agents"
        ],
        "operationId": "archive_agent",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "agents"
        ],
        "operationId": "update_agent",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent_id}/config": {
      "get": {
        "tags": [
          "agents"
        ],
        "operationId": "get_agent_config",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentConfig"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "agents"
        ],
        "operationId": "update_agent_config",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAgentConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentConfig"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agents/{agent_id}/provision": {
      "post": {
        "tags": [
          "agents"
        ],
        "operationId": "provision_agent",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent_id}/restart": {
      "post": {
        "tags": [
          "agents"
        ],
        "operationId": "restart_agent",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeOperationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent_id}/skills": {
      "get": {
        "tags": [
          "agents"
        ],
        "operationId": "list_agent_skills",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentSkill"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent_id}/skills/{skill_name}": {
      "put": {
        "tags": [
          "agents"
        ],
        "operationId": "update_agent_skill",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "skill_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSkillRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSkill"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent_id}/start": {
      "post": {
        "tags": [
          "agents"
        ],
        "operationId": "start_agent",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeOperationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent_id}/stop": {
      "post": {
        "tags": [
          "agents"
        ],
        "operationId": "stop_agent",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeOperationResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Agent": {
        "type": "object",
        "required": [
          "id",
          "ordinal",
          "name",
          "kind",
          "product_role",
          "role",
          "status",
          "display_name",
          "paths",
          "runtime",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "api_port": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "created_at": {
            "$ref": "#/components/schemas/String"
          },
          "dashboard_port": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "display_name": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "$ref": "#/components/schemas/AgentKind"
          },
          "name": {
            "type": "string"
          },
          "namespace_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "ordinal": {
            "type": "integer",
            "format": "int32"
          },
          "paths": {
            "$ref": "#/components/schemas/AgentPaths"
          },
          "product_role": {
            "$ref": "#/components/schemas/AgentProductRole"
          },
          "role": {
            "$ref": "#/components/schemas/AgentRole"
          },
          "runtime": {
            "$ref": "#/components/schemas/AgentRuntime"
          },
          "runtime_version": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/AgentStatus"
          },
          "updated_at": {
            "$ref": "#/components/schemas/String"
          },
          "workflow_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "AgentConfig": {
        "type": "object",
        "required": [
          "agent_id",
          "config_json",
          "soul_md",
          "env_json",
          "updated_at"
        ],
        "properties": {
          "agent_id": {
            "type": "string",
            "format": "uuid"
          },
          "config_json": {},
          "env_json": {},
          "soul_md": {
            "type": "string"
          },
          "updated_at": {
            "$ref": "#/components/schemas/String"
          }
        }
      },
...
      "AgentKind": {
        "type": "string",
        "enum": [
          "hermes",
          "java_agent"
        ]
      },
...
      "AgentPaths": {
        "type": "object",
        "required": [
          "runtime",
          "config",
          "workspace",
          "logs"
        ],
        "properties": {
          "config": {
            "type": "string"
          },
          "logs": {
            "type": "string"
          },
          "runtime": {
            "type": "string"
          },
          "workspace": {
            "type": "string"
          }
        }
      },
      "AgentProductRole": {
        "type": "string",
        "enum": [
          "leader",
          "executor"
        ]
      },
      "AgentRole": {
        "type": "string",
        "enum": [
          "developer",
          "tester",
          "it_lead",
          "custom"
        ]
      },
      "AgentRuntime": {
        "type": "object",
        "required": [
          "desired_state",
          "command_preview",
          "env_preview",
          "last_capabilities_json"
        ],
        "properties": {
          "command_preview": {
            "type": "string"
          },
          "desired_state": {
            "$ref": "#/components/schemas/DesiredState"
          },
          "env_preview": {},
          "health_detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "health_status": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_capabilities_json": {},
          "last_health_at": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/String"
              }
            ]
          },
          "pid": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "started_at": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/String"
              }
            ]
          },
          "startup_command_redacted": {
            "type": [
              "string",
              "null"
            ]
          },
          "stopped_at": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/String"
              }
            ]
          }
        }
      },
...
      "AgentSkill": {
        "type": "object",
        "required": [
          "id",
          "agent_id",
          "name",
          "title",
          "state",
          "source",
          "updated_at"
        ],
        "properties": {
          "agent_id": {
            "type": "string",
            "format": "uuid"
          },
          "content": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/SkillState"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "$ref": "#/components/schemas/String"
          }
        }
      },
      "AgentStatus": {
        "type": "string",
        "enum": [
          "provisioning",
          "ready",
          "starting",
          "running",
          "degraded",
          "stopped",
          "failed",
          "archived"
        ]
      },
...
      "CreateAgentRequest": {
        "type": "object",
        "required": [
          "kind",
          "role",
          "display_name"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "display_name": {
            "type": "string"
          },
          "executor_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "kind": {
            "$ref": "#/components/schemas/AgentKind"
          },
          "namespace_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "namespace_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "product_role": {
            "$ref": "#/components/schemas/AgentProductRole"
          },
          "role": {
            "$ref": "#/components/schemas/AgentRole"
          },
          "workflow_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "workflow_name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "DesiredState": {
        "type": "string",
        "enum": [
          "running",
          "stopped"
        ]
      },
...
      "RuntimeOperationResponse": {
        "type": "object",
        "required": [
          "agent_id",
          "status",
          "message"
        ],
        "properties": {
          "agent_id": {
            "type": "string",
            "format": "uuid"
          },
          "message": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AgentStatus"
          }
        }
      },
...
      "SkillState": {
        "type": "string",
        "enum": [
          "enabled",
          "disabled",
          "missing",
          "dirty"
        ]
      },
...
      "String": {
        "type": "string"
      },
...
      "UpdateAgentConfigRequest": {
        "type": "object",
        "required": [
          "config_json",
          "soul_md",
          "env_json"
        ],
        "properties": {
          "config_json": {},
          "env_json": {},
          "soul_md": {
            "type": "string"
          }
        }
      },
      "UpdateAgentRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "executor_ids": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "namespace_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "product_role": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AgentProductRole"
              }
            ]
          },
          "role": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AgentRole"
              }
            ]
          },
          "workflow_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "UpdateSkillRequest": {
        "type": "object",
        "required": [
          "state"
        ],
        "properties": {
          "content": {
            "type": [
              "string",
              "null"
            ]
          },
          "state": {
            "$ref": "#/components/schemas/SkillState"
          }
        }
      },
...
    }
  },
...
}